backend/mysite/requirements.txt
adrf==0.1.8
aiocache==0.12.3
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
asgiref==3.8.1
async-property==0.2.2
backoff==2.2.1
backports.zoneinfo==0.2.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
channels==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dataclasses-json==0.6.7
DateTime==5.5
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.14
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
email-validator==2.2.0
entrypoints==0.3
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.14.0
httptools==0.6.1
httpx==0.27.0
idna==2.8
importlib-metadata==8.2.0
importlib-resources==6.4.0
inflect==7.3.1
jinja2==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
more-itertools==10.4.0
mypy-extensions==1.0.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
opencage==2.4.0
packaging==24.1
pandas==2.0.3
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.9
pycairo==1.16.2
pycups==1.9.73
pydantic==2.8.2
pydantic-core==2.20.1
pygltflib==1.16.2
pygments==2.18.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.32.3
requests-unixsocket==0.2.0
rich==13.7.1
SecretStorage==2.3.1
shellingham==1.5.4
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.53
sqlparse==0.5.1
starlette==0.37.2
systemd-python==234
typeguard==4.3.0
typer==0.12.3
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.30.4
uvloop==0.19.0
validate-email==1.3
wadllib==1.3.3
watchfiles==0.22.0
websockets==12.0
whitenoise==6.7.0
wrapt==1.16.0
wxPython==4.0.7
xkit==0.0.0
zipp==3.19.2
zope.interface==7.0
